requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
asyncpg==0.30.0
bcrypt==4.0.1
fastapi[standard]==0.115.11
PyJWT==2.10.1
passlib==1.7.4
pytest-asyncio==0.26.0
flake8==7.2.0
pydantic-settings==2.8.1
tzdata==2025.2
SQLAlchemy==2.0.38